# lock: optimize re-entrant performance