```
julia> r = ReentrantLock(); lock(r); unlock(r);
julia> f(r) = @time for i = 1:10^7; lock(r); unlock(r); end;

before:
julia> f(r)
  0.271201 seconds
julia> lock(r); f(r); unlock(r)
  0.274864 seconds

after:
julia> f(r)
  0.280015 seconds
julia> lock(r); f(r); unlock(r)
  0.149524 seconds
```

This also may need #24431 to show better improvement. Both are also being harmed by incorrect PhiC node placement (will open an issue about this).