@yuyichao 2020-01-01
I think this is currently undefined behavior.