@yuyichao 2020-01-01
The part that's accessing shared field without lock. The read (and the corresponding store) of the `locked_by` field must be atomic (relaxed order).