@vtjnash 2020-01-01
`unlock` usually should contain the memory barrier to make that unnecessary