@yuyichao 2020-01-01
No the barrier always works in pairs. The unlock can have all the barrier you can possibly implement and it won't make the unprotected access in lock any safer.