@vtjnash 2020-07-10
> The read (and the corresponding store) of the locked_by field must be atomic (relaxed order).

Now implemented in #36507